+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [0.1.2] - 2024-01-09
+
 - fix and add doc.
 - add development dependency
 - avoid instance variable name conflict (`@models` to `@__models`)